Struct glium_sdl2_hack::SdlGlWindowBackend

pub struct SdlGlWindowBackend { /* fields omitted */ }

This type is transferrable to another thread.

When acquired the context will already be released so all you can do with it is build Glium (which will automatically re-acquire the context).

Methods

impl SdlGlWindowBackend

pub fn build_glium(self) -> Result<SdlGliumDisplayFacade, IncompatibleOpenGl>

Build Glium with current context checks and with default debug callback behavior.

pub fn build_glium_unchecked(
self
) -> Result<SdlGliumDisplayFacade, IncompatibleOpenGl>

Build Glium without current context checks and with default debug callback behavior.

pub fn build_glium_debug(
self,
debug: DebugCallbackBehavior
) -> Result<SdlGliumDisplayFacade, IncompatibleOpenGl>

Build Glium with current context checks and with the given debug callback behavior.

pub fn build_glium_unchecked_debug(
self,
debug: DebugCallbackBehavior
) -> Result<SdlGliumDisplayFacade, IncompatibleOpenGl>

Build Glium without current context checks and with the given debug callback behavior.