Struct glium_sdl2_hack::SdlGliumDisplayFacade

pub struct SdlGliumDisplayFacade { /* fields omitted */ }

Represents the combined SDL window context + Glium rendering context.

⚠ Warning: while this type allows references to the sdl2::video::Window type (through the unsafe window and window_mut methods), and from this a reference to the sdl2::VideoSubsystem is possible, it does not contain a "real" reference to this type since it is not safe to send to another thread. This is enforced because it is not safe to create a window on a thread other than the main thread, so the VideoSubsystem::window function must not be called.

TODO: since we already fork sdl2, could we add a global atomic flag to prevent ever trying to build another window after the first ?

Methods

impl SdlGliumDisplayFacade

pub unsafe fn window(&self) -> &Window

⚠ Warning: the returned window reference does not contain a "real" reference to the sdl2::VideoSubsystem. While most methods should work, it is not possible to create a window from a thread other than the main thread, so the sdl2::VideoSubsystem::window function to build a new window should not be called.

pub unsafe fn window_mut(&mut self) -> &mut Window

⚠ Warning: the returned window reference does not contain a "real" reference to the sdl2::VideoSubsystem. While most methods should work, it is not possible to create a window from a thread other than the main thread, so the sdl2::VideoSubsystem::window function to build a new window should not be called.

pub fn draw(&self) -> Frame

Start drawing on the backbuffer.

This function returns a Frame, which can be used to draw on it. When the Frame is destroyed, the buffers are swapped.

Note that destroying a Frame is immediate, even if vsync is enabled.